+++ lib/pitchfork.rb
@@ -33,6 +33,16 @@ module Pitchfork
 
   BootFailure = Class.new(StandardError)
 
+  # Returns true if the server is shutting down.
+  # This can be useful to implement health check endpoints, so they
+  # can fail immediately after TERM/QUIT/INT was received by the master
+  # process.
+  # Otherwise they may succeed while Pitchfork is draining requests causing
+  # more requests to be sent.
+  def self.shutting_down?
+    SharedMemory.shutting_down?
+  end
+
   # :stopdoc:
 
   # This returns a lambda to pass in as the app, this does not "build" the