+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+- Add `Pitchfork.shutting_down?` to allow health check endpoints to fail sooner on graceful shutdowns.
 - Treat `TERM` as graceful shutdown rather than quick shutdown.
 - Implement `after_worker_hard_timeout` callback.